fix bugs

+++ api/account_financial_year_op.go
@@ -12,11 +12,11 @@ func NewAccountFinancialYearOpService(c *Client) *AccountFinancialYearOpService 
 	return &AccountFinancialYearOpService{client: c}
 }
 
-func (svc *AccountFinancialYearOpService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountFinancialYearOpService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountFinancialYearOpModel, name)
 }
 
-func (svc *AccountFinancialYearOpService) GetByIds(ids []int) (*types.AccountFinancialYearOps, error) {
+func (svc *AccountFinancialYearOpService) GetByIds(ids []int64) (*types.AccountFinancialYearOps, error) {
 	a := &types.AccountFinancialYearOps{}
 	return a, svc.client.getByIds(types.AccountFinancialYearOpModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountFinancialYearOpService) GetAll() (*types.AccountFinancialYearO
 	return a, svc.client.getAll(types.AccountFinancialYearOpModel, a)
 }
 
-func (svc *AccountFinancialYearOpService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountFinancialYearOpService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountFinancialYearOpModel, fields, relations)
 }
 
-func (svc *AccountFinancialYearOpService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountFinancialYearOpService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountFinancialYearOpModel, ids, fields, relations)
 }
 
-func (svc *AccountFinancialYearOpService) Delete(ids []int) error {
+func (svc *AccountFinancialYearOpService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountFinancialYearOpModel, ids)
 }
