+++ api/account_group.go
@@ -12,11 +12,11 @@ func NewAccountGroupService(c *Client) *AccountGroupService {
 	return &AccountGroupService{client: c}
 }
 
-func (svc *AccountGroupService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountGroupService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountGroupModel, name)
 }
 
-func (svc *AccountGroupService) GetByIds(ids []int) (*types.AccountGroups, error) {
+func (svc *AccountGroupService) GetByIds(ids []int64) (*types.AccountGroups, error) {
 	a := &types.AccountGroups{}
 	return a, svc.client.getByIds(types.AccountGroupModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountGroupService) GetAll() (*types.AccountGroups, error) {
 	return a, svc.client.getAll(types.AccountGroupModel, a)
 }
 
-func (svc *AccountGroupService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountGroupService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountGroupModel, fields, relations)
 }
 
-func (svc *AccountGroupService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountGroupService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountGroupModel, ids, fields, relations)
 }
 
-func (svc *AccountGroupService) Delete(ids []int) error {
+func (svc *AccountGroupService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountGroupModel, ids)
 }
