+++ api/format_address_mixin.go
@@ -12,11 +12,11 @@ func NewFormatAddressMixinService(c *Client) *FormatAddressMixinService {
 	return &FormatAddressMixinService{client: c}
 }
 
-func (svc *FormatAddressMixinService) GetIdsByName(name string) ([]int, error) {
+func (svc *FormatAddressMixinService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.FormatAddressMixinModel, name)
 }
 
-func (svc *FormatAddressMixinService) GetByIds(ids []int) (*types.FormatAddressMixins, error) {
+func (svc *FormatAddressMixinService) GetByIds(ids []int64) (*types.FormatAddressMixins, error) {
 	f := &types.FormatAddressMixins{}
 	return f, svc.client.getByIds(types.FormatAddressMixinModel, ids, f)
 }
@@ -36,14 +36,14 @@ func (svc *FormatAddressMixinService) GetAll() (*types.FormatAddressMixins, erro
 	return f, svc.client.getAll(types.FormatAddressMixinModel, f)
 }
 
-func (svc *FormatAddressMixinService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *FormatAddressMixinService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.FormatAddressMixinModel, fields, relations)
 }
 
-func (svc *FormatAddressMixinService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *FormatAddressMixinService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.FormatAddressMixinModel, ids, fields, relations)
 }
 
-func (svc *FormatAddressMixinService) Delete(ids []int) error {
+func (svc *FormatAddressMixinService) Delete(ids []int64) error {
 	return svc.client.delete(types.FormatAddressMixinModel, ids)
 }
