+++ api/ir_default.go
@@ -12,11 +12,11 @@ func NewIrDefaultService(c *Client) *IrDefaultService {
 	return &IrDefaultService{client: c}
 }
 
-func (svc *IrDefaultService) GetIdsByName(name string) ([]int, error) {
+func (svc *IrDefaultService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.IrDefaultModel, name)
 }
 
-func (svc *IrDefaultService) GetByIds(ids []int) (*types.IrDefaults, error) {
+func (svc *IrDefaultService) GetByIds(ids []int64) (*types.IrDefaults, error) {
 	i := &types.IrDefaults{}
 	return i, svc.client.getByIds(types.IrDefaultModel, ids, i)
 }
@@ -36,14 +36,14 @@ func (svc *IrDefaultService) GetAll() (*types.IrDefaults, error) {
 	return i, svc.client.getAll(types.IrDefaultModel, i)
 }
 
-func (svc *IrDefaultService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *IrDefaultService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.IrDefaultModel, fields, relations)
 }
 
-func (svc *IrDefaultService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *IrDefaultService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.IrDefaultModel, ids, fields, relations)
 }
 
-func (svc *IrDefaultService) Delete(ids []int) error {
+func (svc *IrDefaultService) Delete(ids []int64) error {
 	return svc.client.delete(types.IrDefaultModel, ids)
 }
