+++ api/ir_module_module_exclusion.go
@@ -12,11 +12,11 @@ func NewIrModuleModuleExclusionService(c *Client) *IrModuleModuleExclusionServic
 	return &IrModuleModuleExclusionService{client: c}
 }
 
-func (svc *IrModuleModuleExclusionService) GetIdsByName(name string) ([]int, error) {
+func (svc *IrModuleModuleExclusionService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.IrModuleModuleExclusionModel, name)
 }
 
-func (svc *IrModuleModuleExclusionService) GetByIds(ids []int) (*types.IrModuleModuleExclusions, error) {
+func (svc *IrModuleModuleExclusionService) GetByIds(ids []int64) (*types.IrModuleModuleExclusions, error) {
 	i := &types.IrModuleModuleExclusions{}
 	return i, svc.client.getByIds(types.IrModuleModuleExclusionModel, ids, i)
 }
@@ -36,14 +36,14 @@ func (svc *IrModuleModuleExclusionService) GetAll() (*types.IrModuleModuleExclus
 	return i, svc.client.getAll(types.IrModuleModuleExclusionModel, i)
 }
 
-func (svc *IrModuleModuleExclusionService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *IrModuleModuleExclusionService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.IrModuleModuleExclusionModel, fields, relations)
 }
 
-func (svc *IrModuleModuleExclusionService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *IrModuleModuleExclusionService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.IrModuleModuleExclusionModel, ids, fields, relations)
 }
 
-func (svc *IrModuleModuleExclusionService) Delete(ids []int) error {
+func (svc *IrModuleModuleExclusionService) Delete(ids []int64) error {
 	return svc.client.delete(types.IrModuleModuleExclusionModel, ids)
 }
