+++ api/payment_icon.go
@@ -12,11 +12,11 @@ func NewPaymentIconService(c *Client) *PaymentIconService {
 	return &PaymentIconService{client: c}
 }
 
-func (svc *PaymentIconService) GetIdsByName(name string) ([]int, error) {
+func (svc *PaymentIconService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.PaymentIconModel, name)
 }
 
-func (svc *PaymentIconService) GetByIds(ids []int) (*types.PaymentIcons, error) {
+func (svc *PaymentIconService) GetByIds(ids []int64) (*types.PaymentIcons, error) {
 	p := &types.PaymentIcons{}
 	return p, svc.client.getByIds(types.PaymentIconModel, ids, p)
 }
@@ -36,14 +36,14 @@ func (svc *PaymentIconService) GetAll() (*types.PaymentIcons, error) {
 	return p, svc.client.getAll(types.PaymentIconModel, p)
 }
 
-func (svc *PaymentIconService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *PaymentIconService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.PaymentIconModel, fields, relations)
 }
 
-func (svc *PaymentIconService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *PaymentIconService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.PaymentIconModel, ids, fields, relations)
 }
 
-func (svc *PaymentIconService) Delete(ids []int) error {
+func (svc *PaymentIconService) Delete(ids []int64) error {
 	return svc.client.delete(types.PaymentIconModel, ids)
 }
