+++ api/print_prenumbered_checks.go
@@ -12,11 +12,11 @@ func NewPrintPrenumberedChecksService(c *Client) *PrintPrenumberedChecksService 
 	return &PrintPrenumberedChecksService{client: c}
 }
 
-func (svc *PrintPrenumberedChecksService) GetIdsByName(name string) ([]int, error) {
+func (svc *PrintPrenumberedChecksService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.PrintPrenumberedChecksModel, name)
 }
 
-func (svc *PrintPrenumberedChecksService) GetByIds(ids []int) (*types.PrintPrenumberedCheckss, error) {
+func (svc *PrintPrenumberedChecksService) GetByIds(ids []int64) (*types.PrintPrenumberedCheckss, error) {
 	p := &types.PrintPrenumberedCheckss{}
 	return p, svc.client.getByIds(types.PrintPrenumberedChecksModel, ids, p)
 }
@@ -36,14 +36,14 @@ func (svc *PrintPrenumberedChecksService) GetAll() (*types.PrintPrenumberedCheck
 	return p, svc.client.getAll(types.PrintPrenumberedChecksModel, p)
 }
 
-func (svc *PrintPrenumberedChecksService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *PrintPrenumberedChecksService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.PrintPrenumberedChecksModel, fields, relations)
 }
 
-func (svc *PrintPrenumberedChecksService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *PrintPrenumberedChecksService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.PrintPrenumberedChecksModel, ids, fields, relations)
 }
 
-func (svc *PrintPrenumberedChecksService) Delete(ids []int) error {
+func (svc *PrintPrenumberedChecksService) Delete(ids []int64) error {
 	return svc.client.delete(types.PrintPrenumberedChecksModel, ids)
 }
