+++ api/res_partner_industry.go
@@ -12,11 +12,11 @@ func NewResPartnerIndustryService(c *Client) *ResPartnerIndustryService {
 	return &ResPartnerIndustryService{client: c}
 }
 
-func (svc *ResPartnerIndustryService) GetIdsByName(name string) ([]int, error) {
+func (svc *ResPartnerIndustryService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.ResPartnerIndustryModel, name)
 }
 
-func (svc *ResPartnerIndustryService) GetByIds(ids []int) (*types.ResPartnerIndustrys, error) {
+func (svc *ResPartnerIndustryService) GetByIds(ids []int64) (*types.ResPartnerIndustrys, error) {
 	r := &types.ResPartnerIndustrys{}
 	return r, svc.client.getByIds(types.ResPartnerIndustryModel, ids, r)
 }
@@ -36,14 +36,14 @@ func (svc *ResPartnerIndustryService) GetAll() (*types.ResPartnerIndustrys, erro
 	return r, svc.client.getAll(types.ResPartnerIndustryModel, r)
 }
 
-func (svc *ResPartnerIndustryService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *ResPartnerIndustryService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.ResPartnerIndustryModel, fields, relations)
 }
 
-func (svc *ResPartnerIndustryService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *ResPartnerIndustryService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.ResPartnerIndustryModel, ids, fields, relations)
 }
 
-func (svc *ResPartnerIndustryService) Delete(ids []int) error {
+func (svc *ResPartnerIndustryService) Delete(ids []int64) error {
 	return svc.client.delete(types.ResPartnerIndustryModel, ids)
 }
