+++ api/resource_mixin.go
@@ -12,11 +12,11 @@ func NewResourceMixinService(c *Client) *ResourceMixinService {
 	return &ResourceMixinService{client: c}
 }
 
-func (svc *ResourceMixinService) GetIdsByName(name string) ([]int, error) {
+func (svc *ResourceMixinService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.ResourceMixinModel, name)
 }
 
-func (svc *ResourceMixinService) GetByIds(ids []int) (*types.ResourceMixins, error) {
+func (svc *ResourceMixinService) GetByIds(ids []int64) (*types.ResourceMixins, error) {
 	r := &types.ResourceMixins{}
 	return r, svc.client.getByIds(types.ResourceMixinModel, ids, r)
 }
@@ -36,14 +36,14 @@ func (svc *ResourceMixinService) GetAll() (*types.ResourceMixins, error) {
 	return r, svc.client.getAll(types.ResourceMixinModel, r)
 }
 
-func (svc *ResourceMixinService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *ResourceMixinService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.ResourceMixinModel, fields, relations)
 }
 
-func (svc *ResourceMixinService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *ResourceMixinService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.ResourceMixinModel, ids, fields, relations)
 }
 
-func (svc *ResourceMixinService) Delete(ids []int) error {
+func (svc *ResourceMixinService) Delete(ids []int64) error {
 	return svc.client.delete(types.ResourceMixinModel, ids)
 }
