+++ api/resource_test.go
@@ -12,11 +12,11 @@ func NewResourceTestService(c *Client) *ResourceTestService {
 	return &ResourceTestService{client: c}
 }
 
-func (svc *ResourceTestService) GetIdsByName(name string) ([]int, error) {
+func (svc *ResourceTestService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.ResourceTestModel, name)
 }
 
-func (svc *ResourceTestService) GetByIds(ids []int) (*types.ResourceTests, error) {
+func (svc *ResourceTestService) GetByIds(ids []int64) (*types.ResourceTests, error) {
 	r := &types.ResourceTests{}
 	return r, svc.client.getByIds(types.ResourceTestModel, ids, r)
 }
@@ -36,14 +36,14 @@ func (svc *ResourceTestService) GetAll() (*types.ResourceTests, error) {
 	return r, svc.client.getAll(types.ResourceTestModel, r)
 }
 
-func (svc *ResourceTestService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *ResourceTestService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.ResourceTestModel, fields, relations)
 }
 
-func (svc *ResourceTestService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *ResourceTestService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.ResourceTestModel, ids, fields, relations)
 }
 
-func (svc *ResourceTestService) Delete(ids []int) error {
+func (svc *ResourceTestService) Delete(ids []int64) error {
 	return svc.client.delete(types.ResourceTestModel, ids)
 }
